agent-worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-worker"
version = "0.1.0"
description = "Agente financiero mínimo en Python usando LangGraph y Gemini"
requires-python = ">=3.11"
dependencies = [
    "langgraph",
    "langgraph-checkpoint-postgres",
    "langchain-google-genai",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "psycopg[binary]",
    "psycopg-pool",
    "httpx",
]

[tool.hatch.build.targets.wheel]
packages = ["worker"]